\name{hofbatting}
\alias{hofbatting}

\title{
  Hall of Fame Batting Statistics
}
\description{
 Career batting statistics for players inducted in the baseball Hall of Fame
}
\usage{
  hofbatting
}
\format{
   A data frame with 156 observations on the following 8 variables.
  \describe{
\item{Player}{player name}
\item{HR}{career home runs}
\item{AB}{career at-bats}
\item{From}{first season in MLB}
\item{To}{last season in MLB}
\item{OBP}{career on-base percentage}
\item{SLG}{career slugging percentage}
\item{OPS}{career OPS statistic}
  }
}
\source{
Baseball Reference.
}
\keyword{datasets}
