\name{hofpitching}
\alias{hofpitching}

\title{
  Hall of Fame Pitching Statistics
}
\description{
 Career pitching statistics for players inducted in the baseball Hall of Fame
}
\usage{
  hofpitching
}
\format{
   A data frame with 70 observations on the following 30 variables.
  \describe{
\item{Rk}{}
\item{X2}{player name}
\item{Inducted}{year inducated in HOF}
\item{Yrs}{length of career}
\item{From}{first year in MLB}
\item{To}{final year in MLB}
\item{ASG}{number of all-star game appearances}
\item{WAR}{career WAR measure}
\item{W}{career wins}
\item{L}{career losses}
\item{W.L.}{winning fraction}
\item{ERA}{career ERA}
\item{G}{career games}
\item{GS}{career games started}
\item{GF}{career games finished}
\item{CG}{complete games}
\item{SHO}{shutouts}
\item{SV}{saves}
\item{IP}{innings pitched}
\item{H}{hits allowed}
\item{R}{runs allowed}
\item{ER}{earned runs allowed}
\item{HR}{home runs allowed}
\item{BB}{walks}
\item{IBB}{intentional walks}
\item{SO}{strikeouts}
\item{HBP}{hit by pitches}
\item{BK}{balks}
\item{WP}{wild pitches}
\item{BF}{batters faced}
  }
}
\source{
Baseball Reference.
}
\keyword{datasets}
