\name{NLbatting}
\alias{NLbatting}

\title{
  National League Batting Statistics
}
\description{
 Batting statistics for all National League teams in the 2011 season
}
\usage{
  NLbatting
}
\format{
   A data frame with 16 observations on the following 28 variables.
  \describe{
\item{Tm}{team abbreviation}
\item{NBat}{number of players}
\item{BatAge}{average age}
\item{R.G}{runs scored per game}
\item{G}{games played}
\item{PA}{plate appearances}
\item{AB}{at-bats}
\item{R}{runs scored}
\item{H}{hits}
\item{X2B}{doubles}
\item{X3B}{triples}
\item{HR}{home runs}
\item{RBI}{runs batted in}
\item{SB}{stolen bases}
\item{CS}{caught stealing}
\item{BB}{walks}
\item{SO}{strikeouts}
\item{BA}{batting average}
\item{OBP}{on-base percentage}
\item{SLG}{slugging percentage}
\item{OPS}{on-base percentage plus slugging percentage}
\item{OPS.}{relative OPS statistic}
\item{TB}{total bases}
\item{GDP}{grounded into double play}
\item{HBP}{hit by pitch}
\item{SH}{sacrifice hits}
\item{SF}{sacrifice flies}
\item{IBB}{intentional walks}
  }
}
\source{
Baseball Reference.
}
\keyword{datasets}
