\name{NLpitching}
\alias{NLpitching}

\title{
  National League Pitching Statistics
}
\description{
 Pitching statistics for all National League teams in the 2011 season
}
\usage{
  NLpitching
}
\format{
   A data frame with 16 observations on the following 34 variables.
  \describe{
\item{Tm}{team abbreviation}
\item{NPitch}{number of pitchers}
\item{PitchAge}{average age}
\item{R.G}{runs allowed per game}
\item{W}{wins}
\item{L}{losses}
\item{W.L.}{winning fraction}
\item{ERA}{earned run average}
\item{G}{games played}
\item{GS}{games started}
\item{GF}{games finished}
\item{CG}{complete games}
\item{tmSHO}{team shutouts}
\item{cgSHO}{complete game shutouts}
\item{SV}{saves}
\item{IP}{innings pitched}
\item{H}{hits}
\item{R}{runs allowed}
\item{ER}{earned runs allowed}
\item{HR}{home runs}
\item{BB}{walks}
\item{IBB}{intentional walks}
\item{SO}{strikeouts}
\item{HBP}{hit by pitch}
\item{BK}{balks}
\item{WP}{wild pitches}
\item{BF}{batters faced}
\item{ERA.}{relative ERA}
\item{WHIP}{hits and walks per nine innings}
\item{H.9}{hits per nine innings}
\item{HR.9}{home runs per nine innings}
\item{BB.9}{walks per nine innings}
\item{SO.9}{strikeouts per nine innings}
\item{SO.BB}{ratio of strikeouts to walks}
  }
}
\source{
Baseball Reference.
}
\keyword{datasets}
