\name{williams_1941}
\alias{williams_1941}

\title{
  Ted Williams 1941 Batting Data
}
\description{
 Game-by-game hitting data for Ted Williams in the 1941 season
}
\usage{
  williams_1941
}
\format{
   A data frame with 143 observations on the following 33 variables.
  \describe{
 \item{Rk}{}
 \item{Gcar}{}
 \item{Gtm}{game number}
 \item{Date}{date of game}
 \item{Tm}{player team}
 \item{X6}{indicates home or awy}
 \item{Opp}{opposing team}
 \item{Rslt}{game outcome}
 \item{Inngs}{number of innings played}
 \item{PA}{plate appearances}
 \item{AB}{at-bats}
 \item{R}{runs scored}
\item{H}{hits}
\item{X2B}{doubles}
\item{X3B}{triples}
\item{HR}{home runs}
\item{RBI}{runs batted in}
\item{BB}{walks}
\item{IBB}{intentional walks}
\item{SO}{strikeouts}
\item{HBP}{hit by pitches}
\item{SH}{sacrifice hits}
\item{SF}{sacrifice flies}
\item{ROE}{reached on an error}
\item{GDP}{grounded in double play}
\item{SB}{stolen bases}
\item{CS}{caught stealing}
\item{BA}{batting average}
\item{OBP}{on-base percentage}
\item{SLG}{slugging percentage}
\item{OPS}{ops statistic}
\item{BOP}{base-out percentage}
\item{Pos}{fielding position}
  }
}
\source{
Baseball Reference.
}
\keyword{datasets}
